--- a/sql/core/src/main/scala/org/apache/spark/sql/execution/ExpandExec.scala
+++ b/sql/core/src/main/scala/org/apache/spark/sql/execution/ExpandExec.scala
@@ -169,7 +169,7 @@ case class ExpandExec(
 
     // Part 2: switch/case statements
     initBlock += ctx.subexpressionElimination(
-      projections.flatten.map(BindReferences.bindReference(_, attributeSeq)): _*)
+      projections.flatten.map(BindReferences.bindReference(_, attributeSeq)))
     val switchCaseExprs = projections.zipWithIndex.map { case (exprs, row) =>
       val (exprCodesWithIndices, inputVarSets) = exprs.indices.flatMap { col =>
         if (!sameOutput(col)) {