… sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/CodeGenerationSuite.scala
… sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/SubexpressionEliminationSuite.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/aggregate/AggregateCodegenSupport.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/aggregate/HashAggregateExec.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/basicPhysicalOperators.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/joins/BroadcastNestedLoopJoinExec.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/joins/JoinCodegenSupport.scala
… sql/core/src/test/scala/org/apache/spark/sql/execution/WholeStageCodegenSuite.scala